reviews_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
requests